Add config file to publish repo as DRAFT
This JSON file will be parsed to convert the adoc via Asciidoctor and publish the DRAFT files in HTML and PDF.

You can add or remove documents from being published by adding or removing the nodes. For example, if you do not want to publish the guide, just remove the section titled "2".  Or you can add an extension by creating a node "3", etc. 
 "a" are attributes that use the "-a" flag.
 "r" are required modules or gems that use the "-r" flag.
 "path" is relative to the root of the repo.